backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
transformers==4.35.2
torch==2.1.1
yt-dlp==2023.12.30
Pillow==10.2.0
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0 